# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FichoChat API",
        "description": "API REST versionnée de FichoChat. Chaque endpoint documente des exemples JSON complets de réponses (succès et erreurs).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/admin/stats": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tableau de bord admin (compteurs uniquement)",
                "description": "Aucun contenu de message / média / clé. Compteurs agrégés pour supervision.",
                "operationId": "adminStats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminStats200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister les utilisateurs",
                "operationId": "adminUsersIndex",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUsers200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Détail utilisateur (métadonnées)",
                "operationId": "adminUsersShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserShow200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/suspend": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Suspendre un utilisateur",
                "description": "Révoque toutes les sessions. Impossible sur soi-même ou un autre admin.",
                "operationId": "adminUsersSuspend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Abuse report",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Suspendu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserAction200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}/activate": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Réactiver un utilisateur",
                "operationId": "adminUsersActivate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réactivé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminUserAction200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/devices": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lister les appareils (métadonnées)",
                "operationId": "adminDevicesIndex",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDevices200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/devices/{id}/revoke": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Révoquer un appareil (force admin)",
                "operationId": "adminDevicesRevoke",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Révoqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDeviceRevoke200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/security/events": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Journaux de sécurité (global)",
                "operationId": "adminSecurityEvents",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 40
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminSecurityEvents200"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès admin requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorForbidden403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorForbidden403": {
                "description": "Accès refusé (email non vérifié, etc.)",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Email address is not verified."
                    }
                },
                "type": "object",
                "example": {
                    "success": false,
                    "message": "Email address is not verified."
                }
            },
            "ErrorUnauthorized401": {
                "description": "Non authentifié",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object",
                "example": {
                    "success": false,
                    "message": "Unauthenticated."
                }
            },
...
            "AdminStats200": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Admin stats ready.",
                    "data": {
                        "users": {
                            "total": 10,
                            "active": 8,
                            "pending": 1,
                            "suspended": 1,
                            "admins": 1
                        },
                        "devices": {
                            "total": 15,
                            "approved": 12,
                            "pending": 2,
                            "revoked": 1
                        },
                        "conversations": 20,
                        "messages": 100,
                        "medias": 5,
                        "security_events_24h": 12,
                        "activity_logs_24h": 40,
                        "login_failures_24h": 2,
                        "generated_at": "2026-08-04T03:30:00.000000Z"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Access Token obtenu via /auth/login ou /auth/register. Header: Authorization: Bearer {access_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}